% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_auc.R
\name{calc_auc}
\alias{calc_auc}
\title{Calculate diagnostic performance metrics}
\usage{
calc_auc(
  actual,
  predicted,
  score = NULL,
  ci_method_auc = "delong",
  alpha = 0.05,
  boot_n = 2000
)
}
\arguments{
\item{actual}{Vectors of actual values(0=negative, 1=positive)}

\item{predicted}{Vector of predicted binary labels}

\item{score}{Optional probability scores for AUC}

\item{alpha}{type I error}

\item{boot_n}{The number of Bootstrap sampling}

\item{ci_method}{Select method to calculate CI for AUC}
}
\value{
Data frame which contains calculated auc and result of pROC::roc
}
\description{
Calculate diagnostic performance metrics
}
