% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_binary_metrics.R
\name{calc_binary_metrics}
\alias{calc_binary_metrics}
\title{Calculate binary performance metrics}
\usage{
calc_binary_metrics(
  actual,
  predicted,
  metrics_to_calc = c("sensitivity", "specificity", "ppv", "npv", "accuracy"),
  ci_method = "cp",
  alpha = 0.05
)
}
\arguments{
\item{actual}{Vectors of actual values(0=negative, 1=positive)}

\item{predicted}{Vector of predicted binary labels}

\item{metrics_to_calc}{Character vector of metrics to calculate}

\item{ci_method}{Select method to calculate CI}

\item{alpha}{type I error}
}
\value{
A named list of data frame which contains calculated performance metrics and 2X2 confusion matrix table
}
\description{
Calculate binary performance metrics
}
